Move exponential histogram mapping functions into top-level  package
