+++ data/exponential/mapping/exponent/exponent.go
@@ -12,14 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package exponent // import "go.opentelemetry.io/otel/sdk/metric/aggregator/exponential/mapping/exponent"
+package exponent // import "go.opentelemetry.io/otel/data/exponential/mapping/exponent"
 
 import (
 	"fmt"
 	"math"
 
-	"go.opentelemetry.io/otel/sdk/metric/aggregator/exponential/mapping"
-	"go.opentelemetry.io/otel/sdk/metric/aggregator/exponential/mapping/internal"
+	"go.opentelemetry.io/otel/data/exponential/mapping"
+	"go.opentelemetry.io/otel/data/exponential/mapping/internal"
 )
 
 const (
