+++ data/exponential/mapping/internal/float64.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package internal // import "go.opentelemetry.io/otel/sdk/metric/aggregator/exponential/mapping/internal"
+package internal // import "go.opentelemetry.io/otel/data/exponential/mapping/internal"
 
 import "math"
