+++ data/exponential/mapping/mapping.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package mapping // import "go.opentelemetry.io/otel/sdk/metric/aggregator/exponential/mapping"
+package mapping // import "go.opentelemetry.io/otel/data/exponential/mapping"
 
 import "fmt"
